Create Map Page
Retrieved an API key from Google and modified the TiApp.xml to enable
the map module on Apple and Android. Created a map.js page with a
simple map that opens up at Five Guys in San Jose. Not sure if having a
separate js is the correct way so this may change.